Fix all specs regarding abilities
