+++ app/components/saved_scenarios/info/component.html.erb
