+++ app/components/saved_scenarios/info_users/component.rb
@@ -1,11 +1,11 @@
-module SavedScenarioInfoUsers
+module SavedScenarios::InfoUsers
   class Component < ApplicationComponent
     option :users
     option :title
 
     # Initials to show
     def initials_for(saved_scenario_user)
-      saved_scenario_user.user_email.first.capitalize
+      saved_scenario_user.initials.capitalize
     end
   end
 end
