+++ app/components/saved_scenarios/nav_item/component.html.erb
