+++ app/components/saved_scenarios/publish/component.html.erb
