+++ app/components/saved_scenarios/row/component.html.erb
