+++ app/controllers/identity/identity_controller.rb
@@ -5,7 +5,7 @@ module Identity
     extend ActiveSupport::Concern
 
     included do
-      layout 'identity'
+      # layout 'identity'
       before_action :authenticate_user!
       before_action :set_back_url
     end
