+++ app/models/saved_scenario.rb
@@ -16,7 +16,7 @@ class SavedScenario < ApplicationRecord
   has_one :featured_scenario, dependent: :destroy
   has_many :saved_scenario_users, dependent: :destroy
   has_many :users, through: :saved_scenario_users
-  # has_many :users, through: :saved_scenario_users
+
   has_rich_text :description
 
   validates :scenario_id, presence: true
@@ -50,5 +50,29 @@ class SavedScenario < ApplicationRecord
 
   def days_until_last_update
     (Time.current - updated_at) / 60 / 60 / 24
+  end
+
+  def self.owned_by?(user)
+    joins(:saved_scenario_users)
+      .where(
+        'saved_scenario_users.user_id': user.id,
+        'saved_scenario_users.role_id': User::Roles.index_of(:scenario_owner)
+      )
+  end
+
+  def self.collaborated_by?(user)
+    joins(:saved_scenario_users)
+      .where(
+        'saved_scenario_users.user_id': user.id,
+        'saved_scenario_users.role_id': User::Roles.index_of(:scenario_collaborator)..
+      )
+  end
+
+  def self.viewable_by?(user)
+    joins(:saved_scenario_users)
+      .where(
+        'saved_scenario_users.user_id': user.id,
+        'saved_scenario_users.role_id': User::Roles.index_of(:scenario_viewer)..
+      )
   end
 end
