+++ app/views/identity/_sidebar.html.erb
@@ -1,4 +1,4 @@
-<% content_for :sidebar do %>
+<% content_for :block_right do %>
   <%= render(Identity::SidebarItemComponent.new(
     path: identity_profile_path,
     title: t('identity.settings.index.title'),
