+++ app/views/layouts/_sidebar.html.haml
@@ -15,5 +15,5 @@
     .p-5.pl-6.text-midnight-800
       %span= heroicon 'language', options: { class: 'inline-block w-6 h-6' }
       Switch language
-    .bg-midnight-600.py-3
-      = render(SidebarItem::Component.new(path: saved_scenarios_path, title: t('sidebar.profile'), icon: 'user-circle', active: controller_name == 'user', text: 'text-midnight-800'))
+    .bg-midnight-600.py-3{ class: 'hover:underline' }
+      = render(SidebarItem::ProfileComponent.new(path: identity_profile_path, title: t('sidebar.profile'), icon: 'user-circle', active: controller_name == 'settings', text: 'text-midnight-800'))
