+++ app/views/layouts/errors.html.haml
@@ -1,0 +1,28 @@
+!!! 5
+%html{:lang => 'en'}
+  %head
+    %title= content_for(:title) || "ETM"
+    %meta{ name: "viewport", content: "width=device-width,initial-scale=1"}
+    %meta{ name: "apple-mobile-web-app-capable", content: "yes"}
+    = csrf_meta_tags
+    = csp_meta_tag
+
+    = yield :head
+
+    %link{ rel:"manifest", href:"/manifest.json" }
+
+    = stylesheet_link_tag "tailwind", "inter-font", "data-turbo-track": "reload"
+    = stylesheet_link_tag "application", "data-turbo-track": "reload"
+
+    = javascript_importmap_tags
+  %body
+    - if current_user
+      = render partial: "layouts/sidebar"
+      .p-10.py-5.flex.flex-col.h-full{class: 'ml-[300px]'}
+        = render partial: "layouts/buttons"
+        .text-xl.mb-5.mt-5
+          = t('errors.sorry')
+        .grow
+          .flex.h-full
+            %div{class: 'basis-3/4'}
+              = yield
