+++ app/views/saved_scenarios/_block_right_menu.html.erb
@@ -1,58 +1,60 @@
-<%= render(SavedScenarioNavItem::Component.new(
-  path: saved_scenario_path(saved_scenario),
-  title: t('scenario_bar.info'),
-  icon: 'information-circle',
-  active: controller_name == 'saved_scenarios' && action_name == 'show')
-)%>
-<%= render(SavedScenarioNavItem::Component.new(
-  path: saved_scenario_path(saved_scenario),
-  title: t('scenario_bar.history'),
-  icon: 'clock',
-  active: controller_name == 'history')
-)%>
-<%= render(SavedScenarioNavItem::Component.new(
-  path: saved_scenario_path(saved_scenario),
-  title: t('scenario_bar.manage_access'),
-  icon: 'user-group',
-  active: controller_name == 'saved_scenario_users')
-)%>
+<% content_for :block_right do %>
+  <%= render(SavedScenarios::NavItem::Component.new(
+    path: saved_scenario_path(@saved_scenario),
+    title: t('scenario_bar.info'),
+    icon: 'information-circle',
+    active: controller_name == 'saved_scenarios' && action_name == 'show')
+  )%>
+  <%= render(SavedScenarios::NavItem::Component.new(
+    path: saved_scenario_path(@saved_scenario),
+    title: t('scenario_bar.history'),
+    icon: 'clock',
+    active: controller_name == 'history')
+  )%>
+  <%= render(SavedScenarios::NavItem::Component.new(
+    path: saved_scenario_path(@saved_scenario),
+    title: t('scenario_bar.manage_access'),
+    icon: 'user-group',
+    active: controller_name == 'saved_scenario_users')
+  )%>
 
-<div class="border-t border-solid border-gray-200 my-10"></div>
+  <div class="border-t border-solid border-gray-200 my-10"></div>
 
-<%= render(Hovercard::Component.new(
-    path: '',
-    text: t("scenario_bar.private.description.#{saved_scenario.private}")
-  )) do %>
-  <%= render(PublishSavedScenario::Component.new(
-    path_on: publish_saved_scenario_path(saved_scenario),
-    path_off: unpublish_saved_scenario_path(saved_scenario),
-    status: saved_scenario.private,
-    title: t("scenario_bar.private.#{saved_scenario.private}"),
-    icon_on:'eye-slash',
-    icon_off: 'eye',
-  ))%>
+  <%= render(Hovercard::Component.new(
+      path: '',
+      text: t("scenario_bar.private.description.#{@saved_scenario.private}")
+    )) do %>
+    <%= render(SavedScenarios::Publish::Component.new(
+      path_on: publish_saved_scenario_path(@saved_scenario),
+      path_off: unpublish_saved_scenario_path(@saved_scenario),
+      status: @saved_scenario.private,
+      title: t("scenario_bar.private.#{@saved_scenario.private}"),
+      icon_on:'eye-slash',
+      icon_off: 'eye',
+    ))%>
+  <% end %>
+  <%= render(SavedScenarios::NavItem::Component.new(
+    path: saved_scenario_path(@saved_scenario),
+    title: t("scenario_bar.featured.#{@saved_scenario.featured?}"),
+    icon: @saved_scenario.featured? ? 'star' : 'sparkles',
+    static: true)
+  )%>
+  <%= render(Hovercard::Component.new(
+      path: '',
+      text: t("scenario_bar.edit.description")
+    )) do %>
+    <%= render(SavedScenarios::NavItem::Component.new(
+      path: edit_saved_scenario_path(@saved_scenario),
+      title: t("scenario_bar.edit.title"),
+      icon: 'pencil',
+      static: true,
+      active: action_name == 'edit'
+    ))%>
+  <% end %>
+  <%= render(SavedScenarios::NavItem::Component.new(
+    path: saved_scenario_path(@saved_scenario),
+    title: t("scenario_bar.discard"),
+    icon: 'trash',
+    static: true)
+  )%>
 <% end %>
-<%= render(SavedScenarioNavItem::Component.new(
-  path: saved_scenario_path(saved_scenario),
-  title: t("scenario_bar.featured.#{saved_scenario.featured?}"),
-  icon: saved_scenario.featured? ? 'star' : 'sparkles',
-  static: true)
-)%>
-<%= render(Hovercard::Component.new(
-    path: '',
-    text: t("scenario_bar.edit.description")
-  )) do %>
-  <%= render(SavedScenarioNavItem::Component.new(
-    path: edit_saved_scenario_path(saved_scenario),
-    title: t("scenario_bar.edit.title"),
-    icon: 'pencil',
-    static: true,
-    active: action_name == 'edit'
-  ))%>
-<% end %>
-<%= render(SavedScenarioNavItem::Component.new(
-  path: saved_scenario_path(saved_scenario),
-  title: t("scenario_bar.discard"),
-  icon: 'trash',
-  static: true)
-)%>
