+++ config/locales/en.yml
@@ -93,3 +93,5 @@ en:
     no_description: This scenario has no description. Click here to create one.
   saved_scenarios:
     empty: You don't have any scenarios.
+  flash:
+    need_login: Please log in again
